// src/app/services/openapi-client/model/brandDto.ts
/**
 * PaintDepotApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BrandDto { 
    id?: number;
    name?: string | null;
    slug?: string | null;
    logoUrl?: string | null;
    description?: string | null;
    websiteUrl?: string | null;
    isActive?: boolean;
    displayOrder?: number;
    productCount?: number;
    createdAt?: string | null;
}