// src/app/services/openapi-client/model/createBrandDto.ts
/**
 * PaintDepotApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateBrandDto { 
    name: string;
    slug: string;
    logoUrl?: string | null;
    description?: string | null;
    websiteUrl?: string | null;
    isActive?: boolean;
    displayOrder?: number;
}